EFRS-671: All integration tests use real database
