+++ pom.xml
@@ -6,7 +6,7 @@
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.2.5.RELEASE</version>
+        <version>2.2.8.RELEASE</version>
         <relativePath/> <!-- lookup parent from repository -->
     </parent>
     <groupId>com.exadel</groupId>
@@ -20,20 +20,19 @@
     </modules>
 
     <properties>
-        <spring.boot.version>2.2.5.RELEASE</spring.boot.version>
-        <spring.oauth2.jwt.version>1.0.10.RELEASE</spring.oauth2.jwt.version>
+        <spring.boot.version>2.2.8.RELEASE</spring.boot.version>
+        <spring.oauth2.jwt.version>1.1.1.RELEASE</spring.oauth2.jwt.version>
         <java.version>11</java.version>
-        <mapstruct.version>1.3.0.Final</mapstruct.version>
-        <lombok.version>1.18.8</lombok.version>
+        <mapstruct.version>1.3.1.Final</mapstruct.version>
+        <lombok.version>1.18.12</lombok.version>
         <swagger.version>2.9.2</swagger.version>
         <docker.image.prefix>frs</docker.image.prefix>
         <smile.version>2.4.0</smile.version>
         <io.jsonwebtoken.version>0.9.1</io.jsonwebtoken.version>
-        <config.typesafe.version>1.4.0</config.typesafe.version>
         <feign-form.version>3.8.0</feign-form.version>
         <feign-core.version>10.7.4</feign-core.version>
-        <commons-lang3.version>3.9</commons-lang3.version>
-        <spring-cloud-starter-openfeign.version>2.2.2.RELEASE</spring-cloud-starter-openfeign.version>
+        <commons-lang3.version>3.11</commons-lang3.version>
+        <spring-cloud-starter-openfeign.version>2.2.3.RELEASE</spring-cloud-starter-openfeign.version>
         <ehcache.version>2.7.2</ehcache.version>
         <javax-cache.version>1.1.1</javax-cache.version>
         <logstash-logback-encoder.version>5.3</logstash-logback-encoder.version>
@@ -94,11 +93,6 @@
                 <groupId>io.jsonwebtoken</groupId>
                 <artifactId>jjwt</artifactId>
                 <version>${io.jsonwebtoken.version}</version>
-            </dependency>
-            <dependency>
-                <groupId>com.typesafe</groupId>
-                <artifactId>config</artifactId>
-                <version>${config.typesafe.version}</version>
             </dependency>
             <dependency>
                 <groupId>org.springframework.cloud</groupId>
